requirements.txt
# Data Processing
pandas==2.0.3
numpy==1.24.3
faker==18.13.0

# Database
sqlalchemy==2.0.19
psycopg2-binary==2.9.6

# Machine Learning
scikit-learn==1.3.0
xgboost==1.7.6
imbalanced-learn==0.11.0

# API
flask==2.3.2
flask-cors==4.0.0
gunicorn==21.2.0

# Orchestration
apache-airflow==2.6.3
apache-airflow-providers-postgres==5.5.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Development
jupyter==1.0.0
pytest==7.4.0
black==23.7.0
flake8==6.1.0

# Big Data (optional)
pyspark==3.4.1

# Monitoring
mlflow==2.5.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0